refactor(domain): GameRule 기반 랭크 판단 수정
serice에서 기존에 rank에서 판단하던 로직을 gameRule에서 판단을 진행하도록 수정하였습니다.

이를 통해 Lottos 클래스에 GameRule을 만들 객체를 전달하여 Lottos 클래스에서 해당 룰에 맞는 판단을 진행하도록 구현하였습니다.
