+++ src/main/java/lotto/domain/service/LottoService.java
@@ -3,7 +3,8 @@ package lotto.domain.service;
 import lotto.domain.model.Lottos;
 import lotto.domain.model.PurchasePrice;
 import lotto.domain.model.WinningLotto;
-import lotto.domain.result.RankResult;
+import lotto.domain.rule.GameRule;
+import lotto.domain.rule.RankResult;
 import lotto.global.util.Parser;
 
 public class LottoService {
@@ -26,8 +27,9 @@ public class LottoService {
         return lottoFactory.createWinningLotto(parser.parse(winningLotto), Integer.parseInt(bonusLotto));
     }
 
-    // 당첨 로또와 비교 후 매칭카운트, 보너스 히트 판별 각 로또에 Rank 저장
+    // 당첨 로또와 비교 후 매칭카운트, 보너스 히트 판별
     public RankResult judgeLottosRanks(Lottos lottos, WinningLotto winningLotto) {
-        return lottos.judgeLottosWithWinningLotto(winningLotto);
+        GameRule lottoGameRule = GameRuleFactory.createLottoGameRule();
+        return lottos.judgeLottosWithWinningLotto(winningLotto, lottoGameRule);
     }
 }
